fix reviewed issues
